+++ lithops/job/job.py
@@ -38,13 +38,27 @@ logger = logging.getLogger(__name__)
 FUNCTION_CACHE = set()
 
 
-def create_map_job(config, internal_storage, executor_id, job_id, map_function,
-                   iterdata,  runtime_meta, runtime_memory, extra_env,
-                   include_modules, exclude_modules, execution_timeout,
-                   chunksize=None, extra_args=None, obj_chunk_size=None,
-                   obj_chunk_number=None):
+def create_map_job(
+    config, 
+    internal_storage,
+    executor_id,
+    job_id, 
+    map_function,
+    iterdata,
+    runtime_meta,
+    runtime_memory,
+    extra_env,
+    include_modules,
+    exclude_modules,
+    execution_timeout,
+    chunksize=None,
+    extra_args=None,
+    obj_chunk_size=None,
+    obj_newline='\n',
+    obj_chunk_number=None
+):
     """
-    Wrapper to create a map job.  It integrates COS logic to process objects.
+    Wrapper to create a map job. It integrates COS logic to process objects.
     """
     host_job_meta = {'host_job_create_tstamp': time.time()}
     map_iterdata = utils.verify_args(map_function, iterdata, extra_args)
@@ -56,27 +70,29 @@ def create_map_job(config, internal_storage, executor_id, job_id, map_function,
         # Create partitions according chunk_size or chunk_number
         logger.debug('ExecutorID {} | JobID {} - Calling map on partitions '
                      'from object storage flow'.format(executor_id, job_id))
-        map_iterdata, ppo = create_partitions(config, internal_storage,
-                                              map_iterdata, obj_chunk_size,
-                                              obj_chunk_number)
-
+        map_iterdata, ppo = create_partitions(
+            config, internal_storage, map_iterdata, 
+            obj_chunk_size, obj_chunk_number, obj_newline
+        )
         host_job_meta['host_job_create_partitions_time'] = round(time.time()-create_partitions_start, 6)
     # ########
 
-    job = _create_job(config=config,
-                      internal_storage=internal_storage,
-                      executor_id=executor_id,
-                      job_id=job_id,
-                      func=map_function,
-                      iterdata=map_iterdata,
-                      chunksize=chunksize,
-                      runtime_meta=runtime_meta,
-                      runtime_memory=runtime_memory,
-                      extra_env=extra_env,
-                      include_modules=include_modules,
-                      exclude_modules=exclude_modules,
-                      execution_timeout=execution_timeout,
-                      host_job_meta=host_job_meta)
+    job = _create_job(
+        config=config,
+        internal_storage=internal_storage,
+        executor_id=executor_id,
+        job_id=job_id,
+        func=map_function,
+        iterdata=map_iterdata,
+        chunksize=chunksize,
+        runtime_meta=runtime_meta,
+        runtime_memory=runtime_memory,
+        extra_env=extra_env,
+        include_modules=include_modules,
+        exclude_modules=exclude_modules,
+        execution_timeout=execution_timeout,
+        host_job_meta=host_job_meta
+    )
 
     if ppo:
         job.parts_per_object = ppo
@@ -84,10 +100,22 @@ def create_map_job(config, internal_storage, executor_id, job_id, map_function,
     return job
 
 
-def create_reduce_job(config, internal_storage, executor_id, reduce_job_id,
-                      reduce_function, map_job, map_futures, runtime_meta,
-                      runtime_memory, reducer_one_per_object, extra_env,
-                      include_modules, exclude_modules, execution_timeout=None):
+def create_reduce_job(
+    config,
+    internal_storage,
+    executor_id,
+    reduce_job_id,
+    reduce_function,
+    map_job,
+    map_futures,
+    runtime_meta,
+    runtime_memory,
+    obj_reduce_by_key,
+    extra_env,
+    include_modules,
+    exclude_modules,
+    execution_timeout=None
+):
     """
     Wrapper to create a reduce job. Apply a function across all map futures.
     """
@@ -95,7 +123,7 @@ def create_reduce_job(config, internal_storage, executor_id, reduce_job_id,
 
     iterdata = [(map_futures, )]
 
-    if hasattr(map_job, 'parts_per_object') and reducer_one_per_object:
+    if hasattr(map_job, 'parts_per_object') and obj_reduce_by_key:
         prev_total_partitons = 0
         iterdata = []
         for total_partitions in map_job.parts_per_object:
@@ -111,25 +139,39 @@ def create_reduce_job(config, internal_storage, executor_id, reduce_job_id,
 
     iterdata = utils.verify_args(reduce_function, iterdata, None)
 
-    return _create_job(config=config,
-                       internal_storage=internal_storage,
-                       executor_id=executor_id,
-                       job_id=reduce_job_id,
-                       func=reduce_function,
-                       iterdata=iterdata,
-                       runtime_meta=runtime_meta,
-                       runtime_memory=runtime_memory,
-                       extra_env=ext_env,
-                       include_modules=include_modules,
-                       exclude_modules=exclude_modules,
-                       execution_timeout=execution_timeout,
-                       host_job_meta=host_job_meta)
+    return _create_job(
+        config=config,
+        internal_storage=internal_storage,
+        executor_id=executor_id,
+        job_id=reduce_job_id,
+        func=reduce_function,
+        iterdata=iterdata,
+        runtime_meta=runtime_meta,
+        runtime_memory=runtime_memory,
+        extra_env=ext_env,
+        include_modules=include_modules,
+        exclude_modules=exclude_modules,
+        execution_timeout=execution_timeout,
+        host_job_meta=host_job_meta
+    )
 
 
-def _create_job(config, internal_storage, executor_id, job_id, func,
-                iterdata,  runtime_meta, runtime_memory, extra_env,
-                include_modules, exclude_modules, execution_timeout,
-                host_job_meta, chunksize=None):
+def _create_job(
+    config,
+    internal_storage,
+    executor_id,
+    job_id,
+    func,
+    iterdata,
+    runtime_meta,
+    runtime_memory,
+    extra_env,
+    include_modules,
+    exclude_modules,
+    execution_timeout,
+    host_job_meta,
+    chunksize=None
+):
     """
     Creates a new Job
     """
@@ -277,7 +319,12 @@ def _create_job(config, internal_storage, executor_id, job_id, func,
     return job
 
 
-def _store_func_and_modules(job_tmp_dir, func_key, func_str, module_data):
+def _store_func_and_modules(
+    job_tmp_dir,
+    func_key,
+    func_str,
+    module_data
+):
     ''' stores function and modules in temporary directory to be
     used later in optimized runtime
     '''