+++ lithops/storage/utils.py
@@ -34,14 +34,13 @@ init_key_suffix = ".init"
 
 class StorageNoSuchKeyError(Exception):
     def __init__(self, bucket, key):
-        msg = "No such key /{}/{} found in storage.".format(bucket, key)
+        msg = f"No such key /{bucket}/{key} found in storage."
         super(StorageNoSuchKeyError, self).__init__(msg)
 
 
 class StorageConfigMismatchError(Exception):
     def __init__(self, current_path, prev_path):
-        msg = "The data is stored at {}, but current storage is configured at {}.".format(
-            prev_path, current_path)
+        msg = f"The data is stored at {prev_path}, but current storage is configured at {current_path}"
         super(StorageConfigMismatchError, self).__init__(msg)
 
 
@@ -52,8 +51,8 @@ class CloudObject:
         self.key = key
 
     def __str__(self):
-        path = '{}://{}/{}'.format(self.backend, self.bucket, self.key)
-        return '<CloudObject at {}>'.format(path)
+        path = f'{self.backend}://{self.bucket}/{self.key}'
+        return f'<CloudObject at {path}>'
 
 
 class CloudObjectUrl:
@@ -61,7 +60,7 @@ class CloudObjectUrl:
         self.url = url
 
     def __str__(self):
-        return '<CloudObject at {}>'.format(self.url)
+        return f'<CloudObject at {self.url}>'
 
 
 class CloudObjectLocal:
@@ -71,7 +70,7 @@ class CloudObjectLocal:
         self.key = os.path.basename(path)
 
     def __str__(self):
-        return '<CloudObject at {}>'.format(self.path)
+        return f'<CloudObject at {self.path}>'
 
 
 def clean_bucket(storage, bucket, prefix, sleep=5):
@@ -79,8 +78,8 @@ def clean_bucket(storage, bucket, prefix, sleep=5):
     Deletes all the files from COS. These files include the function,
     the data serialization and the function invocation results.
     """
-    msg = "Going to delete all objects from bucket '{}'".format(bucket)
-    msg = msg + " and prefix '{}'".format(prefix) if prefix else msg
+    msg = f"Going to delete all objects from bucket '{bucket}'"
+    msg = msg + f" and prefix '{prefix}'" if prefix else msg
     logger.info(msg)
     total_objects = 0
     objects_to_delete = storage.list_keys(bucket, prefix)
@@ -91,7 +90,7 @@ def clean_bucket(storage, bucket, prefix, sleep=5):
         time.sleep(sleep)
         objects_to_delete = storage.list_keys(bucket, prefix)
 
-    logger.info('Finished deleting objects, total found: {}'.format(total_objects))
+    logger.info(f'Finished deleting objects, total found: {total_objects}')
 
 
 def create_job_key(executor_id, job_id):
@@ -162,8 +161,7 @@ def create_init_key(executor_id, job_id, call_id, act_id):
     :return: output key
     """
     job_key = create_job_key(executor_id, job_id)
-    return '/'.join([JOBS_PREFIX, job_key, call_id,
-                     '{}{}'.format(act_id, init_key_suffix)])
+    return '/'.join([JOBS_PREFIX, job_key, call_id, f'{act_id}{init_key_suffix}'])
 
 
 def get_storage_path(storage_config):