+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [Alibaba Cloud Functions] Allow to build a runtime from a custom requirements.txt file
 - [Core] Add support for building container runtimes with podman
 - [Core] Delete only runtimes from the specified backend on 'lithops clean'
+- [Executor] Added obj_newline parameter in map() and map_reduce() methods
 
 ### Changed
 - [Core] Load docker command only when needed instead of loading it always
@@ -20,6 +21,7 @@
 - [Multiprocessing] Improved remote logging
 - [Monitor] Reduce debug log verbosity, status is printed every 30s or when a future changes state
 - [AWS Batch] Increased resource limits
+- [Executor] Changed 'reducer_one_per_object' parameter to 'obj_reduce_by_key'
 
 ### Fixed
 - [Google Cloud Functions] Fixed errors when deploying a function